Allow separate BIDS files with the same sampling frequency but different start times
In the case of manual recording with ideacmdtool in XA30, the recording of the cardiac and respiratory signals does not start or end at the same time. I added code to allow for this case.
I also demean the raw signals to match the pipeline when importing the .puls and .resp files directly.